# openapi.yaml
openapi: 3.1.0
info:
  title: CEOS ARD Editor API
  version: 0.1.0
  description: |
    RESTful API for the CEOS ARD Editor application that provides workspace management,
    file editing, building block management, and GitHub integration capabilities.
  contact:
    name: CEOS ARD Team
    url: https://github.com/ceos-org/ceos-ard-editor
  license:
    name: Apache-2.0
    url: https://github.com/ceos-org/ceos-ard-editor/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://ceos-ard-editor.example.com/api
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /workspaces:
    get:
      summary: List all workspaces
      description: Retrieves all workspaces for the authenticated user
      operationId: listWorkspaces
      tags:
        - Workspaces
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create new workspace
      description: Creates a new workspace with cloned repository
      operationId: createWorkspace
      tags:
        - Workspaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Workspace title
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  description: Workspace description
                pfs:
                  $ref: '#/components/schemas/pfs'
      responses:
        '201':
          description: Workspace created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /workspaces/{workspaceId}/files:
    get:
      summary: List files in workspace
      description: |-
        Retrieves files and folders in the workspace.

        Doesn't list the following files / folders:
        - Files/folders starting with a dot (.)

        Additionalls, doesn't list the following Files/folders at the root:
        - templates
        - build
      operationId: listFiles
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: path
          in: query
          description: Directory path to list (defaults to root)
          schema:
            type: string
            default: "/"
        - name: recurse
          in: query
          description: Whether to list files recursively
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of files and folders
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileItem'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create new empty file or folder
      description: Creates a new empty file or folder in the workspace
      operationId: createFileOrFolder
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: File or folder name
                  minLength: 1
                path:
                  type: string
                  description: Directory path where to create the file/folder
                type:
                  type: string
                  enum:
                    - file
                    - folder
                  description: Type of item to create
              required:
                - name
                - path
                - type
      responses:
        '201':
          description: File created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileItem'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/files/{filePath}:
    get:
      summary: Get file content
      description: |-
        Retrieves the content of a specific file.

        The endoint shall only return files that are made available through
        `GET /workspaces/{workspaceId}/files`. Other files, e.g. withing `.git`
        should not be accessible and return a 404 error.
      operationId: getFileContent
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/FilePath'
      responses:
        '200':
          description: File content
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Upload file content
      description: |-
        Uploads the content of a specific file.

        Validation may take place, e.g. validating against the strictyaml schemas of the corresponding YAML files.
        File should be stored as is without parsing, e.g. comments in YAML files should be preserved.

        The endoint shall only handle files that are made available through
        `GET /workspaces/{workspaceId}/files`. Other files, e.g. withing `.git`
        should not be changed and return a 404 error.
      operationId: uploadFileContent
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/FilePath'
      requestBody:
        required: true
        content:
          '*/*':
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: File updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileItem'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete file
      description: |-
        Deletes a specific file.

        The endoint shall only handle files that are made available through
        `GET /workspaces/{workspaceId}/files`. Other files, e.g. withing `.git`
        should not be deleted and return a 404 error.
      operationId: deleteFile
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/FilePath'
      responses:
        '200':
          description: |-
            A tracked file has been deleted, it could be reverted.
            The file information is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileItem'
        '204':
          description: |-
            An untracked file has been deleted permanently, it can't be restored.
            No file information is returned.
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update file metadata or operations
      description: |-
        Performs file operations such as rename or revert.

        The endoint shall only handle files that are made available through
        `GET /workspaces/{workspaceId}/files`. Other files, e.g. withing `.git`
        should not be changed and return a 404 error.
      operationId: patchFile
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/FilePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operation:
                  type: string
                  enum:
                    - rename
                    - revert
                  description: The operation to perform on the file
              required:
                - operation
              oneOf:
                - properties:
                    operation:
                      const: rename
                    target:
                      description: New name for the file
                      type: string
                  required:
                    - target
                - properties:
                    operation:
                      const: revert
      responses:
        '200':
          description: |-
            File operation completed successfully.
            The updated file information is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileItem'
        '400':
          description: Invalid request data or operation conflicts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/search:
    get:
      summary: Search files in workspace
      description: Searches through all files in the workspace
      operationId: searchFiles
      tags:
        - Workspaces
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: query
          in: query
          required: true
          description: Search query
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      required:
                        - name
                        - path
                        - type
                      properties:
                        name:
                          type: string
                          description: Name to the file containing the match
                        path:
                          type: string
                          description: Absolute path to the file starting from the workspace root
                        type:
                          type: string
                          description: Where the match was found
                          enum:
                            - filename
                            - content
                      oneOf:
                        - type: object
                          properties:
                            type:
                              const: filename
                        - type: object
                          required:
                            - line
                            - column
                            - excerpt
                          properties:
                            type:
                              const: content
                            line:
                              type: integer
                              description: Line number of the match
                            column:
                              type: integer
                              description: Column number of the match
                            excerpt:
                              type: string
                              description: Text excerpt around the match
        '400':
          description: Invalid search query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/diffs:
    get:
      summary: Get list of changed files
      description: Retrieves a list of files that are not unchanged (including added, deleted, modified, etc.) according to git
      operationId: getChangedFiles
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: List of changed files
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      required:
                        - path
                        - status
                      properties:
                        path:
                          type: string
                          description: Absolute path to the file starting from the workspace root
                        status:
                          type: string
                          enum:
                            - A
                            - M
                            - D
                            - R
                          description: |-
                            Status of the file

                            - A=Added
                            - M=Modified
                            - D=Deleted
                            - R=Renamed
                      if:
                        properties:
                          status:
                            const: R
                      then:
                        required:
                          - source
                        properties:
                          source:
                            type: string
                            description: Old path (for renamed files)
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Create a commit of the changes
      description: >-
        Creates a new git commit for all current workspace changes and pushes it to GitHub.

        When the branch on GitHub moved ahead, the remote changes are merged into the workspace
        automatically; if they conflict with the committed changes, a 409 with the conflicting
        files is returned instead and the changes remain staged.
      operationId: commitChanges
      tags:
        - Files
        - Workspaces
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Create a git commit
              required:
                - message
              properties:
                message:
                  $ref: '#/components/schemas/CommitMessage'
      responses:
        '200':
          description: Commit created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitResult'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Remote changes conflict with the committed changes; changes remain staged
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human-readable conflict description
                      conflicting_files:
                        type: array
                        items:
                          type: string
                        description: Paths of the conflicting files
  /workspaces/{workspaceId}/diffs/{filePath}:
    get:
      summary: Get diff for specific file
      description: |-
        Retrieves the git diff for a specific file, if available.

        The endoint shall only return files that are made available through
        `GET /workspaces/{workspaceId}/files`. Other files, e.g. withing `.git`
        should not be accessible and return a 404 error.
      operationId: getFileDiff
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/FilePath'
      responses:
        '200':
          description: File diff information
          content:
            text/plain:
              schema:
                type: string
                description: |-
                  Diff content in unified diff format.
                  Might be empty if there are no changes.
                example: |-
                  diff --git a/glossary/ups.yaml b/glossary/ups.yaml
                  index fafae97..e231b87 100644
                  --- a/glossary/ups.yaml
                  +++ b/glossary/ups.yaml
                  @@ -1,2 +1,2 @@
                  term: UPS
                  -description: Universal Polar Stereographic
                  \ No newline at end of file
                  +description: United Parcel Service
                  \ No newline at end of file
        '404':
          description: Workspace or file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/context/{filePath}:
    get:
      summary: Get additional contextual information for specific file
      description: |-
        Retrieves the context for a specific file, if available.

        This is currently a list of PFS that use the specified file.

        The endoint shall only return files that are made available through
        `GET /workspaces/{workspaceId}/files`. Other files, e.g. withing `.git`
        should not be accessible and return a 404 error.
      operationId: getFileContext
      tags:
        - Files
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/FilePath'
      responses:
        '200':
          description: Contextual information retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/FileItem'
                  - type: object
                    required:
                      - usage
                    properties:
                      usage:
                        description: |-
                          List of PFS IDs that use the specified file.

                          Returns an empty array if the file is not used by any PFS.
                          Returns `null` if the information is not available / unknown.
                        type:
                          - array
                          - 'null'
                        items:
                          type: string
                          description: PFS ID that uses the specified file
        '404':
          description: Workspace or file not found, or no diff available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Unique identifier of the workspace
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required:
        - error
        - code
...
    Workspace:
      type: object
      required:
        - id
        - name
        - description
        - repository
        - branch
        - pfs
        - created
        - pr
        - user_id
      properties:
        id:
          type: string
          format: uuid
          description: Unique workspace identifier
        title:
          type: string
          description: Workspace title
        description:
          type: string
          description: Workspace description
        repository:
          type: string
          format: uri
          description: URL of the Git repository for the workspace
        branch:
          type: string
          description: Name of the branch in the repository
        pfs:
          $ref: '#/components/schemas/pfs'
        created:
          type: string
          format: date-time
          description: Workspace creation timestamp
        status:
          type: string
          enum:
            - active
            - archived
          description: |
            Workspace status.

            - "archived": Workspace has been archived (manually or due to merged/closed PR)
            - "active": Workspace is active and can be worked on
        archived_at:
          type:
            - string
            - "null"
          format: date-time
          description: |
            Timestamp when workspace was archived.

            Set automatically when status changes to "archived".
            Cleared when workspace is reactivated.
          default: null
        deletion_at:
          type:
            - string
            - "null"
          format: date-time
          description: |
            Scheduled permanent deletion timestamp (computed as archived_at + 1 month).

            **Note**: This is a computed property, not stored in the database.
            After this time, workspace files and database record will be permanently deleted by cleanup script.
            Automatically null when workspace is not archived.
          default: null
          readOnly: true
        pr:
          type:
            - string
            - "null"
          description: ID to the GitHub PR
          default: null
        pr_status:
          type:
            - string
            - "null"
          enum:
            - unknown
            - open
            - approved
            - closed
            - merged
            - null
          description: Status of the PR
          default: null
        user_id:
          type: string
          description: The user ID of the workspace owner
...
    pfs:
      description: List of PFS types assigned to the workspace, used for preview generation.
      type: array
      items:
        $ref: '#/components/schemas/pfsId'
      example:
        - NRB
        - POL
...